backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
psycopg[binary]==3.2.1
python-dotenv==1.0.1
